refactoring to shorthand
