+++ tests/index.js
@@ -1,5 +1,5 @@
 module.exports = exports = {
-	"info": require("./info"),
 	"methods": require("./methods"),
 	"types": require("./types"),
+	"util": require("./util"),
 };
