--- tests/info/index.js
@@ -1,6 +1,0 @@
-module.exports = exports = {
-	"createdBy": require("./createdBy"),
-	"created": require("./created"),
-	"modifiedBy": require("./modifiedBy"),
-	"modified": require("./modified"),
-};
