+++ tests/methods/getVersion/static/index.js
@@ -1,5 +1,5 @@
 require('../../../util');
-describe('get version using static method', function(){
+describe('get version using instance method', function(){
 	require('./date');
 	require('./version');
 });
