+++ tests/util/index.js
@@ -1,4 +1,6 @@
 module.exports = exports = {
+	"fixDates": require("./fixDates"),
 	"models": require("./models"),
 	"setup": require("./setup"),
+	"shorthand": require("./shorthand"),
 };
