+++ tests/util/shorthand/index.js
@@ -1,0 +1,6 @@
+module.exports = exports = {
+	"composite": require("./composite"),
+	"compositeWithArray": require("./compositeWithArray"),
+	"reference": require("./reference"),
+	"simple": require("./simple"),
+};
